+++ test/generators/rubocop_install_generator_test.rb
@@ -15,53 +15,77 @@ class RubocopInstallGeneratorTest < Rails::Generators::TestCase
     app_path
   end
 
-  def test_should_install_rubocop_successfully
-    Dir.chdir(app_path) do
-      quietly { run_generator }
+  class BaseConfigurationsTest < self
+    def test_should_install_rubocop_successfully
+      Dir.chdir(app_path) do
+        quietly { run_generator }
 
-      assert_file "Gemfile" do |content|
-        assert_match(/rubocop/, content)
-        assert_match(/rubocop-rails/, content)
-        assert_match(/rubocop-performance/, content)
+        assert_file "Gemfile" do |content|
+          assert_match(/rubocop/, content)
+          assert_match(/rubocop-rails/, content)
+          assert_match(/rubocop-performance/, content)
+          assert_match(/rubocop-rake/, content)
+        end
+
+        assert_file ".rubocop.yml" do |content|
+          assert_match("2.7.1", content)
+          assert_match("AllCops", content)
+        end
       end
+    end
 
-      assert_file ".rubocop.yml" do |content|
-        assert_match("2.7.1", content)
-        assert_match("AllCops", content)
+    def test_should_skip_adding_rubocop_rules
+      Dir.chdir(app_path) do
+        quietly { run_generator [destination_root, "--skip_adding_rubocop_rules"] }
+
+        assert_file ".rubocop.yml" do |content|
+          assert_no_match("AllCops", content)
+          assert_no_match("2.7.1", content)
+        end
+      end
+    end
+
+    def test_should_add_target_ruby_version
+      Dir.chdir(app_path) do
+        quietly { run_generator [destination_root, "--ruby_version=2.6.3"] }
+
+        assert_file ".rubocop.yml" do |content|
+          assert_match("AllCops", content)
+          assert_match("2.6.3", content)
+        end
       end
     end
   end
 
-  def test_should_skip_adding_rubocop_rules
-    Dir.chdir(app_path) do
-      quietly { run_generator [destination_root, "--skip_adding_rubocop_rules"] }
+  class TestFrameworkGemTest < self
+    def test_should_skip_gem_extension
+      quietly { generator.add_rubocop_gems }
 
       assert_file "Gemfile" do |content|
-        assert_match(/rubocop/, content)
-        assert_match(/rubocop-rails/, content)
-        assert_match(/rubocop-performance/, content)
-      end
-
-      assert_file ".rubocop.yml" do |content|
-        assert_no_match("2.7.1", content)
-        assert_no_match("AllCops", content)
+        assert_no_match(/rubocop-rspec/, content)
+        assert_no_match(/rubocop-minitest/, content)
       end
     end
-  end
 
-  def test_should_add_target_ruby_version
-    Dir.chdir(app_path) do
-      quietly { run_generator [destination_root, "--ruby_version=2.6.3"] }
+    def test_should_raise_error_for_invalid_option
+      assert_raises(NotImplementedError) do
+        run_generator ["--test_gem=invalid_test_extension_gem"]
+      end
+    end
+
+    def test_should_configure_correct_gem_extension
+      quietly { run_generator ["--test_gem=minitest"] }
 
       assert_file "Gemfile" do |content|
         assert_match(/rubocop/, content)
         assert_match(/rubocop-rails/, content)
         assert_match(/rubocop-performance/, content)
+        assert_match(/rubocop-rake/, content)
+        assert_match(/rubocop-minitest/, content)
       end
 
       assert_file ".rubocop.yml" do |content|
-        assert_match("2.6.3", content)
-        assert_match("AllCops", content)
+        assert_match("rubocop-minitest", content)
       end
     end
   end